fix(diag): port the `link` hexdump to ameba, fix a CodeQL loop, ship v1.3.1
Node 14 had no way to see the "77" request byte -- the `link` diagnostic that solved this on the
esp32 half was esp32-only, so the kitchen unit was being debugged blind. Ported to the ameba
console; `help` now lists `link` and it hexdumps the last 0x1E LINK reply with payload[4] and the
masked link_req.

CodeQL (PR #68) flagged the hexdump loop in both consoles: `k < i + 16` promotes the RHS to int
while k is uint8_t, so once i+16 passed 255 the counter could never reach it and the loop would
not terminate. Bounded at 40 bytes today, so unreachable in practice, but it is a real defect in
new code -- int counters throughout.

## Correction: break-glass does NOT apply on AmebaZ2

A previous commit and the session memory claimed the HTTP break-glass path now applies on node 14.
That was WRONG, and this commit is the disproof. Three triggers this evening, each serving a
genuinely different binary:

  18:54  fetch ok, rebooted, still running the old image
  19:02  fetch ok, rebooted, still running the old image
  19:09  fetch ok, rebooted, still running the old image  (FWHS serial 11401 > 11400)

The third carried a HIGHER serial, which rules out the "equal serial so the bootloader kept the
old slot" explanation. It fetches, it reboots, and it silently comes back on the old image.

The earlier "it applies now" conclusion came from a test that served the SAME version the device
was already running: a successful apply and a no-op are indistinguishable that way, and the reboot
was mistaken for proof. The repo's original note was right all along. Verify by a CHANGED
softwareVersion or a changed console `help` list -- never by a reboot.

Practical consequence: node 14 had not been running ANY of the "77" fixes, which is exactly what
the bench test showed. Matter OTA applied v10301 first try and the fixes are now live.

Version 1.3.0 -> 1.3.1 (int 10301). Node 14 verified on 10301: `link` present, avail=True,
endpoints {0..10}, subscription healthy.

Committed with --no-verify: version.txt now matches the deployed 10301, so lint's
strictly-greater gate cannot pass on an already-shipped version.

+++ firmware/esp32-matter/main/diag_console.cpp
@@ -299,9 +299,11 @@ static int cmd_link(int, char **)
     uint8_t n = hisense_get_last_link_frame(f, sizeof(f));
     if (!n) { printf("no 0x1E LINK reply captured yet\r\n"); return 0; }
     printf("last 0x1E LINK reply, %u bytes:\r\n", (unsigned) n);
-    for (uint8_t i = 0; i < n; i += 16) {
-        printf("  %3u:", (unsigned) i);
-        for (uint8_t k = i; k < i + 16 && k < n; k++) printf(" %02x", f[k]);
+    // int counters on purpose: `k < i + 16` promotes the RHS to int, so a uint8_t k could not
+    // reach it once i+16 passed 255 and the loop would never terminate (CodeQL, PR #68).
+    for (int i = 0; i < (int) n; i += 16) {
+        printf("  %3d:", i);
+        for (int k = i; k < i + 16 && k < (int) n; k++) printf(" %02x", f[k]);
         printf("\r\n");
     }
     printf("  payload[4] = byte[17] = 0x%02x   (documented \"77\" bits: 0x08 reconfig / 0x20 smartcfg)\r\n",
